+++ lib/ex_unit/test/test_helper.exs
@@ -1,2 +1,3 @@
 # Configure ExUnit, no options supported yet.
 ExUnit.start []
+ExUnit.after_spawn fn -> Process.put(:after_spawn, :ex_unit) end